prepped the app for testing by refactoring index.ts and app.ts and installing dev dependencies to write tests
