+++ auth/Dockerfile
@@ -2,7 +2,7 @@ FROM node:alpine
 
 WORKDIR /app
 COPY package.json .
-RUN npm install
+RUN npm install --only=prod
 COPY . .
 
 CMD ["npm", "start"]
