+++ auth/package.json
@@ -22,5 +22,13 @@
     "mongoose": "^5.11.8",
     "ts-node-dev": "^1.1.1",
     "typescript": "^4.1.3"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.20",
+    "@types/supertest": "^2.0.10",
+    "jest": "^26.6.3",
+    "mongodb-memory-server": "^6.9.2",
+    "supertest": "^6.0.1",
+    "ts-jest": "^26.4.4"
   }
 }
