+++ auth/src/models/user.ts
@@ -56,6 +56,6 @@ userSchema.statics.build = (attrs: UserAttrs) => {
   return new User(attrs);
 };
 
-const User = mongoose.model<UserDoc, UserModel>("User", userSchema);
+const User = mongoose.model<any, any>("User", userSchema);
 
 export { User };
